+++ src/protocol/neoircd.c
@@ -330,7 +330,7 @@ send_burst (int b)
 }
 
 void 
-send_akill (const char *sender, const char *host, const char *ident, const char *setby, const unsigned int length, const char *reason, const unsigned long ts)
+send_akill (const char *sender, const char *host, const char *ident, const char *setby, const unsigned long length, const char *reason, const unsigned long ts)
 {
 	send_cmd (":%s GLINE %s %s %lu :%s", sender, ident, host, (ts + length), reason);
 }
